+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.1.1
 - added check for preceding **gosub** call in **0AA1 ([return_if_false](https://library.sannybuilder.com/#/sa/script/extensions/CLEO/0AA1))**
 - added warnings about exceeding memory allocations count/size limits with **allocate_memory**
+- added listing of remaining memory blocks allocated by scripts to the cleo.log when closing or starting new game
 
 ## 5.1.0
 - fixed collision of GXT texts hook with SAMP and other mods